refactor: Split EHR item into substep components
