+++ src/entities/activity/ui/items/RequestHealthRecordDataItem/OneUpHealthStep.tsx
@@ -1,0 +1,6 @@
+import { FC } from 'react';
+
+// This component is a placeholder as per the original code that had "TODO: iframe"
+export const OneUpHealthStep: FC = () => {
+  return <p>TODO: iframe</p>;
+};
